+++ pymodbus/server/sync.py
@@ -83,6 +83,7 @@ class ModbusBaseRequestHandler(socketserver.BaseRequestHandler):
         """
         raise NotImplementedException("Method not implemented"
                                       " by derived class")
+
     def send(self, message):
         """ Send a request (string) to the network
 
@@ -90,6 +91,7 @@ class ModbusBaseRequestHandler(socketserver.BaseRequestHandler):
         """
         raise NotImplementedException("Method not implemented "
                                       "by derived class")
+
 
 class ModbusSingleRequestHandler(ModbusBaseRequestHandler):
     """ Implements the modbus server protocol
@@ -505,40 +507,54 @@ class ModbusSerialServer(object):
 # --------------------------------------------------------------------------- #
 # Creation Factories
 # --------------------------------------------------------------------------- #
-def StartTcpServer(context=None, identity=None, address=None, **kwargs):
+def StartTcpServer(context=None, identity=None, address=None,
+                   custom_functions=[], **kwargs):
     """ A factory to start and run a tcp modbus server
 
     :param context: The ModbusServerContext datastore
     :param identity: An optional identify structure
     :param address: An optional (interface, port) to bind to.
+    :param custom_functions: An optional list of custom function classes
+        supported by server instance.
     :param ignore_missing_slaves: True to not send errors on a request to a
-                                  missing slave
+                                      missing slave
     """
     framer = kwargs.pop("framer", ModbusSocketFramer)
     server = ModbusTcpServer(context, framer, identity, address, **kwargs)
+
+    for f in custom_functions:
+        server.decoder.register(f)
     server.serve_forever()
 
 
-def StartUdpServer(context=None, identity=None, address=None, **kwargs):
+def StartUdpServer(context=None, identity=None, address=None,
+                   custom_functions=[], **kwargs):
     """ A factory to start and run a udp modbus server
 
     :param context: The ModbusServerContext datastore
     :param identity: An optional identify structure
     :param address: An optional (interface, port) to bind to.
+    :param custom_functions: An optional list of custom function classes
+        supported by server instance.
     :param framer: The framer to operate with (default ModbusSocketFramer)
     :param ignore_missing_slaves: True to not send errors on a request
                                     to a missing slave
     """
     framer = kwargs.pop('framer', ModbusSocketFramer)
     server = ModbusUdpServer(context, framer, identity, address, **kwargs)
+    for f in custom_functions:
+        server.decoder.register(f)
     server.serve_forever()
 
 
-def StartSerialServer(context=None, identity=None, **kwargs):
+def StartSerialServer(context=None, identity=None,  custom_functions=[],
+                      **kwargs):
     """ A factory to start and run a serial modbus server
 
     :param context: The ModbusServerContext datastore
     :param identity: An optional identify structure
+    :param custom_functions: An optional list of custom function classes
+        supported by server instance.
     :param framer: The framer to operate with (default ModbusAsciiFramer)
     :param port: The serial port to attach to
     :param stopbits: The number of stop bits to use
@@ -551,6 +567,8 @@ def StartSerialServer(context=None, identity=None, **kwargs):
     """
     framer = kwargs.pop('framer', ModbusAsciiFramer)
     server = ModbusSerialServer(context, framer, identity, **kwargs)
+    for f in custom_functions:
+        server.decoder.register(f)
     server.serve_forever()
 
 # --------------------------------------------------------------------------- #